Mask more fields in `clean-entry.js`
Also, split out the masking tests to their own file.
